+++ setup.cfg
@@ -47,6 +47,7 @@ install_requires =
     onnxruntime
     skl2onnx
     psyneulink>=0.8.8.1
+    h5py
 
 python_requires = >=3.6
 include_package_data = True
