+++ src/modeci_mdf/interfaces/pytorch/__init__.py
@@ -1,4 +1,4 @@
 from .exporter import pytorch_to_mdf
 
-from .mdf_to_pytorch import mdf_to_pytorch
+from .importer import mdf_to_pytorch
 from . import mod_torch_builtins
